Add function `launch_client` for unified client scripts
This function should make it easier to launch the client in a
standardised way.  It takes as argument a function that initialises and
returns an `Agent` instance.  This agent is then automatically executed,
using server connection information passed either via command line
arguments or environment variables.


## Details

`launch_client` uses argparse to get server connection info (url, port
and access token) from command line arguments or environment variables
(the former preceding the later).  It then calls the given function,
which is expected to return an initialised agent (derived from `Agent`).
Finally the `run()`-method of this agent is called.

In addition to the server info, further arguments can be passed using
`--args` (example: `--args --foo=42 -b`).  Those arguments will be
passed to the initialization function, so it is possible to use them
(e.g. via another argparse instance) to parametrise the agent.

So the expected signature of the decorated function is
```
def foobar(args: list[str]) -> Agent: ...
```

A usage example is provided in `examples/hockey/client_launcher.py`.