Move Python package into sub-dir "comprl"
To have the repository a bit more organised, move the Python code into a
subdirectory "comprl".  With this, the comprl-web, presentation and
documentation directories are now on the same level, outside of the
package directory.
This should make it a bit easier to handle the repo, e.g. when grepping
files or the like.