Use database models from comprl instead of implementing them in reflex
I want to use the data models implemented in comprl, so we don't
replicate stuff in the reflex package.  Unfortunately, this doesn't
integrated well with the way how reflex works.
I think I got it working now, though.  Instead of `rx.session()`, I'm
using vanilla SQLAlchemy for the queries.
One key thing I had to do was to add `MappedAsDataclass` to the
sqlalchemy base class, so that it works with reflex's state
serialisation.