Expect single database with fixed table names
Previous changes (linking between tables) already requires the use of a
single database and table names are not configurable anymore.

The original idea of using separate databases was, to allow easier
backup of the individual tables.  However, I think this is actually not
really needed (you can simply  back them up together).  Having both
tables in the same database allows to use joining on the database level,
which makes the code much nicer.

Further, with the introduction of sqlalchemy, it is not really feasible
anymore to make table names configurable.  However, I also don't think
that this is a feature that is really needed, so simply drop that
option.

Both of the above is already required for the integration with the
Reflex web interface.  However, there was still some code infrastructure
for the configurability, which I am now removing in this commit.

- Remove `ConnectionInfo`.  Since the table name is not configurable
  anymore, just passing the db path directly is enough.
- Add entry "database_path" in ConfigProvider which replaces the
  separate entries for user and games database.
- Replace the separate user/game database arguments with a unified one
  and remove the ones for table names everywhere.
- Update example config files accordingly.