Use variconf for config loading
Using variconf has several advantages:
- Easy merging of values from config file and command line arguments
- Automatic checks for missing values and wrong types
- Using a dataclass for the config instead of the dictionary-like
  ConfigProvider allows autocompletion of config values.

Remove the ConfigProvider class and instead add a module
`server.config`, which holds a `Config` dataclass and a global instance
which can be accessed via `get_config()`.  `load_config()` is used to
load config from the given config file and merges it with potential
overwrites from the command line.

BREAKING: The command line overwrites are now done like this:
```
... --config-overwrites port=1234 data_dir=/foo/bar ...
```
So there is no separate argument for each value anymore.  While this
reduces the ease of use a bit (no help text for possible arguments), it
greatly improves maintainability as config values only need to be
specified in one place.