comprl: New script manage_users.py
Add a unified script for user management.  It provides commands to list,
show, add and edit users.  It replaces the old new_user.py and
update_user.py which are removed.

The script uses typer for command line parsing.  Simply run with
`--help` to show usage information.

It is also installed as executable `comprl-users`, so it can be directly
called with that name (instead of `python -m
comprl.scripts.manage_users`).